lua/lsp.lua: Disable pyflakes when working on anything using nose2
As I called out in the comment, the only work project I have where
nose2 is used is (still, ugh) based on Python 2.7. Since
python-lsp-server/pylsp is Python 3.7+, the copy of pyflakes that it
imports is also running under Python 3.7, and since pyflakes judges
target code based on the version of Python it is invoked with, this
means that it complains about Python 2 syntax in this project.

As a kludge, let's just disable pyflakes if it looks like there is a
`nose2.cfg` file in the working directory of the pylsp server.
It... does the job.